--- src/main/java/codesquad/web/CartController.java
@@ -1,17 +1,0 @@
-package codesquad.web;
-
-import codesquad.service.CartService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class CartController {
-    @Autowired
-    private CartService cartService;
-
-    @GetMapping("/cart")
-    public String show() {
-        return "cart";
-    }
-}
